pyproject.toml
[project]
name = "mips-ae-pylib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.3",
    "inquirerpy>=0.3.4",
    "pycryptodome>=3.23.0",
    "pyelftools>=0.32",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[project.optional-dependencies]
reporting = [
    "pydantic>=2.7",
    "rich>=13.7",
    "pyyaml>=6.0",
    "plotly>=5.20",
    "jinja2>=3.1"
]
server = [
    "fastapi>=0.111",
    "uvicorn>=0.29"
]
notebook = [
    "pandas>=2.2",
    "jupyter"
]
all = [
    "mips-ae-pylib[reporting,server,notebook]"
]

[project.scripts]
atlasexplorer-report = "atlasexplorer.reporting.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
